add fake data

+++ BusinessObjects/Models/GoldPrice.cs
@@ -7,8 +7,8 @@ public partial class GoldPrice
 {
     public int GoldPriceId { get; set; }
     public string? City { get; set; }
-    public decimal BuyPrice { get; set; }
-    public decimal SellPrice { get; set; }
+    public float BuyPrice { get; set; }
+    public float SellPrice { get; set; }
     public string? Type { get; set; }
     
     public DateTime LastUpdated { get; set; }
