+++ BusinessObjects/Models/StonePrice.cs
@@ -4,8 +4,8 @@ public class StonePrice
 {
     public int StonePriceId { get; set; }
     public string? City { get; set; }
-    public decimal BuyPrice { get; set; }
-    public decimal SellPrice { get; set; }
+    public float BuyPrice { get; set; }
+    public float SellPrice { get; set; }
     public string? Type { get; set; }
     public DateTime LastUpdated { get; set; }
